Created Basis for On Rails Movement
Using Bezier Curves, the basis of the on rails movement was created. The routing script should be used for the empty "Route" game objects and these are responsible for keeping track of the different routes that the avatar will follow. Make sure to place the waypoint prefabs under the correct route game objects too to keep things organized. When placing the waypoints for the routes, make sure you adjust the size of the array of waypoints and then drag and drop each one you placed (IN NUMERICAL ORDER!!!!) into the inspector so the curve can be created properly. If done right, the Bezier Curve should be visualized by grey spheres and Moving said waypoints will adjust it. After the routes are complete, click the vr avatar and find the newly placed BezierFollow script in the inspector. Adjust the size of the routes array for how many you want to have and then place the route game objects that contain the waypoints into each slot (ALSO IN ORDER). If done right, the avatar will automatically move along the paths visualized. I already created two routes so you can see it in action here.