Fade To Black and Fade To Clear at the start of the game.
Created two separate scripts; Fade to Black, Fade to Clear. These scripts create a fade to black before the player enters the painting and then fades back to clear when the player is in the painting. The fade to black and fade to clear are activated through a timer.

ToDo:
Combine the scripts into one.